requirements.txt
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.8.3
charset-normalizer==3.4.4
distro==1.9.0
django==4.2.25
django-cors-headers==4.4.0
djangorestframework==3.15.2
exceptiongroup==1.3.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.1
openai==2.1.0
packaging==25.0
pillow==10.4.0
pydantic==2.10.6
pydantic-core==2.27.2
python-dotenv==1.0.1
requests==2.32.4
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-extensions==4.13.2
urllib3==2.2.3
